Create Swap Nodes in Pairs RECURSIVE SOLUTION
Given a linked list, swap every two adjacent nodes and return its head. You may not modify the values in the list's nodes, only nodes itself may be changed. This is the counterpart of the solution with similar name found in this repo, which explores a possible solution using recursion.